# packages/www/assets/openapi.json
{"openapi":"3.0.3","info":{"version":"4.4.2","title":"fastify-swagger"},"components":{"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer"}},"schemas":{}},"paths":{"/api/settings":{"get":{"operationId":"settings","responses":{"200":{"description":"Default Response"}}}},"/api/character/":{"get":{"operationId":"characterGetOne","parameters":[{"in":"query","name":"entry","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"properties":{"sub":{"type":"array","items":{"type":"string"}},"sup":{"type":"array","items":{"type":"string"}},"var":{"type":"array","items":{"type":"string"}},"reading":{"type":"array","items":{"type":"object","additionalProperties":false,"properties":{"type":{"type":"string"},"kana":{"type":"string"}},"required":["kana"]}},"english":{"type":"array","items":{"type":"string"}}},"required":["sub","sup","var","reading","english"],"$schema":"http://json-schema.org/draft-07/schema#"}}}}}}},"/api/character/q":{"get":{"operationId":"characterQuery","parameters":[{"in":"query","name":"q","required":true,"schema":{"type":"string"}},{"in":"query","name":"limit","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"properties":{"result":{"type":"array","items":{"type":"string"}}},"required":["result"],"$schema":"http://json-schema.org/draft-07/schema#"}}}}}}},"/api/character/random":{"get":{"operationId":"characterRandom","responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"properties":{"result":{"type":"string"},"english":{"type":"string"},"level":{"type":"integer"}},"required":["result","english","level"],"$schema":"http://json-schema.org/draft-07/schema#"}}}}}}},"/api/extra/q":{"get":{"operationId":"extraQuery","parameters":[{"in":"query","name":"q","required":true,"schema":{"type":"string"}},{"in":"query","name":"page","required":false,"schema":{"type":"integer"}},{"in":"query","name":"limit","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"properties":{"result":{"type":"array","items":{"type":"object","additionalProperties":false,"properties":{"entry":{"type":"array","items":{"type":"string"}},"reading":{"type":"array","items":{"type":"object","additionalProperties":false,"properties":{"type":{"type":"string"},"kana":{"type":"string"}},"required":["kana"]}},"english":{"type":"array","items":{"type":"string"}}},"required":["entry","reading","english"]}}},"required":["result"],"$schema":"http://json-schema.org/draft-07/schema#"}}}}}}},"/api/sentence/":{"get":{"operationId":"sentenceGetOne","parameters":[{"in":"query","name":"entry","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"properties":{"ja":{"type":"string"},"en":{"type":"string"}},"required":["ja","en"],"$schema":"http://json-schema.org/draft-07/schema#"}}}}}}},"/api/sentence/q":{"get":{"operationId":"sentenceQuery","parameters":[{"in":"query","name":"q","required":false,"schema":{"type":"string"}},{"in":"query","name":"limit","required":false,"schema":{"type":"integer"}},{"in":"query","name":"page","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"properties":{"result":{"type":"array","items":{"type":"object","additionalProperties":false,"properties":{"ja":{"type":"string"},"en":{"type":"string"}},"required":["ja","en"]}}},"required":["result"],"$schema":"http://json-schema.org/draft-07/schema#"}}}}}}},"/api/sentence/random":{"get":{"operationId":"sentenceRandom","responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"properties":{"result":{"type":"string"},"english":{"type":"string"},"level":{"type":"integer"}},"required":["result","english","level"],"$schema":"http://json-schema.org/draft-07/schema#"}}}}}}},"/api/util/tokenize":{"get":{"operationId":"utilTokenize","parameters":[{"in":"query","name":"q","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"properties":{"result":{"type":"array","items":{"type":"object","additionalProperties":true,"properties":{"surface_form":{"type":"string"}},"required":["surface_form"]}}},"required":["result"],"$schema":"http://json-schema.org/draft-07/schema#"}}}}}}},"/api/util/reading":{"get":{"operationId":"utilReading","parameters":[{"in":"query","name":"q","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"properties":{"result":{"type":"string"}},"required":["result"],"$schema":"http://json-schema.org/draft-07/schema#"}}}}}}},"/api/util/speak":{"get":{"operationId":"utilSpeak","parameters":[{"in":"query","name":"q","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Default Response"}}}},"/api/vocabulary/":{"get":{"operationId":"vocabularyGetOne","parameters":[{"in":"query","name":"entry","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"properties":{"entry":{"type":"array","items":{"type":"string"}},"reading":{"type":"array","items":{"type":"object","additionalProperties":false,"properties":{"type":{"type":"string"},"kana":{"type":"string"}},"required":["kana"]}},"english":{"type":"array","items":{"type":"string"}}},"required":["entry","reading","english"],"$schema":"http://json-schema.org/draft-07/schema#"}}}}}}},"/api/vocabulary/q":{"get":{"operationId":"vocabularyQuery","parameters":[{"in":"query","name":"q","required":true,"schema":{"type":"string"}},{"in":"query","name":"limit","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"properties":{"result":{"type":"array","items":{"type":"object","additionalProperties":false,"properties":{"entry":{"type":"array","items":{"type":"string"}},"reading":{"type":"array","items":{"type":"object","additionalProperties":false,"properties":{"type":{"type":"string"},"kana":{"type":"string"}},"required":["kana"]}},"english":{"type":"array","items":{"type":"string"}}},"required":["entry","reading","english"]}}},"required":["result"],"$schema":"http://json-schema.org/draft-07/schema#"}}}}}}},"/api/vocabulary/random":{"get":{"operationId":"vocabularyRandom","responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"properties":{"result":{"type":"string"},"english":{"type":"string"},"level":{"type":"integer"}},"required":["result","english","level"],"$schema":"http://json-schema.org/draft-07/schema#"}}}}}}}},"security":[{"BearerAuth":[]}]}